new file: builds an R package if it's been updated
User specifies the directory of the source package, and the required output
directory. The script determines the name of the output .tar.gz file for the
package.

If the output file does not exist, or if any of the files in the source-tree
for the package are more recent than the output file, the package is built and
put into the .tar.gz file.