add checkout method to ExternalRepository
Permits user to checkout a specific commit for a given repository once it has
been cloned. The program is killed if any exceptions are thrown, eg, if the
requested commit doesn't define a valid commit in the repository.

A unit test was added that checks that appropriate git commands are called by
python. But, this test only replicates the implementation of repo.checkout().

TODO: add integration tests that check that an error occurs when checking out
an invalid commit, and that checking out a valid commit leaves the file
structure in the correct state.

run_workflow (in setup_git_clones) was updated to ensure that repo.checkout()
is called.