ensure PKGNAME is not unbound
PKGNAME is not set in `job_specific_vars.sh` for a project that
does not require a job-specific R package.

The shell scripts fail when unbound variables are dereferenced
(due to set -u)

When called in a project that doesn't require the building of a
project-specific R package, PKGNAME should never be referred to
in setup_libs.sh.